Improved recommending algorythm. It doesn't work as intended yet

+++ recommending_v2/model/constraint.py
@@ -1,7 +1,7 @@
 from typing import List
 from recommending_v2.model.point_of_interest import PointOfInterest
 
-from recommending_v2.categories.categories_graph import CategoriesProvider
+from recommending_v2.categories.categories_provider import CategoriesProvider
 
 
 class Constraint:
@@ -33,7 +33,7 @@ class CategoryConstraint(Constraint):
         self.provider: CategoriesProvider = CategoriesProvider()
         self.weight = 20
 
-    def evaluate(self, poi: PointOfInterest) -> int:
+    def evaluate(self, poi: PointOfInterest) -> float:
         return self.provider.compute_score(poi.kinds, self.codes)
 
     def get_weight(self):
